requirements.txt
Flask
Werkzeug
pytesseract
Pillow
pdf2image
openai
python-dotenv
Flask-Limiter
Flask-Login
bcrypt
gunicorn
